Windows Desktop Tool/requirements.txt
PyQt5
requests
psutil
PyQt-Fluent-Widgets
speedtest-cli
Pillow
pdf2docx
docx2pdf
python-docx
pandas
openpyxl
pywin32
